Expose /root exports containing the createRoot function
