+++ workspaces/rcx/src/index.ts
@@ -1,6 +1,6 @@
 export * from './components/index.ts';
 export * from './context/index.ts';
 export * from './hooks/index.ts';
-export * from './render.ts';
+export * from './root.ts';
 export * from './types.ts';
 export * from './utils/index.ts';
