+++ workspaces/rcx/src/render.ts
@@ -1,9 +1,7 @@
 import { renderingContext } from './components/canvas/context.ts';
-import { emitter } from './internal/emitter.ts';
 import { cxGlobal } from './internal/global.ts';
 import type {
   AnyObject,
-  CreateRootResult,
   NestedArray,
   RCXChild,
   RCXChildren,
@@ -12,17 +10,6 @@ import type {
   RCXRenderingContext,
 } from './types.ts';
 import { isArray } from './utils/type-guards.ts';
-
-const getCanvasElement = (container: HTMLElement) => {
-  if (container instanceof HTMLCanvasElement) {
-    return container;
-  }
-
-  const canvasElement = document.createElement('canvas');
-  container.appendChild(canvasElement);
-
-  return canvasElement;
-};
 
 const unmountNodes = (
   children: NestedArray<RCXNodeAny> | RCXNodeAny | undefined
@@ -143,7 +130,7 @@ const getMatchingChildNext = (
   }
 };
 
-const renderElement = (
+export const renderElement = (
   element: RCXElementAny,
   renderingContextState: RCXRenderingContext,
   prevNode: RCXNodeAny | undefined,
@@ -244,70 +231,4 @@ const renderElement = (
   });
 
   return node;
-};
-
-export const createRoot = (container: HTMLElement): CreateRootResult => {
-  const canvas = getCanvasElement(container);
-  const ctx2d = canvas.getContext('2d');
-
-  if (!ctx2d) {
-    const errorMessage = 'CanvasRenderingContext2D not supported';
-
-    if (window.console && typeof window.console.error === 'function') {
-      // eslint-disable-next-line no-console
-      console.error(errorMessage);
-    }
-
-    return {
-      error: errorMessage,
-    };
-  }
-
-  const renderingContextState = { canvas, ctx2d };
-
-  let rootElement: RCXElementAny | undefined;
-  let raf: number | undefined;
-  let rootNode: RCXNodeAny | undefined;
-
-  const renderRoot = () => {
-    if (typeof raf === 'number') {
-      window.cancelAnimationFrame(raf);
-    }
-
-    raf = window.requestAnimationFrame(() => {
-      // eslint-disable-next-line no-self-assign
-      canvas.width = canvas.width;
-      if (rootElement) {
-        rootNode = renderElement(rootElement, renderingContextState, rootNode);
-      } else {
-        rootNode = undefined;
-      }
-    });
-  };
-
-  const unmount = () => {
-    emitter.off('render', renderRoot);
-    rootElement = undefined;
-
-    if (typeof raf === 'number') {
-      window.cancelAnimationFrame(raf);
-    }
-
-    // eslint-disable-next-line no-self-assign
-    canvas.width = canvas.width;
-
-    if (!(container instanceof HTMLCanvasElement)) {
-      container.removeChild(canvas);
-    }
-  };
-
-  return {
-    render: (element: RCXElementAny) => {
-      rootElement = element;
-      renderRoot();
-
-      emitter.on('render', renderRoot);
-    },
-    unmount,
-  };
 };
